Fixed a bug in the Dialogue System
A bug that if I created two instances of the Dialogue Trigger. When I execute any NPC Script with the Dialogue Trigger having it referenced.

It will only execute the very last instance of the Dialogue Trigger.

So what I did was, I made the Dialogue Data into static variable.